@@ -7,94 +7,86 @@
     "not ie <= 8"
   ],
   "dependencies": {
-    "@types/classnames": "^0.0.32",
-    "@types/history": "^2.0.0",
-    "@types/js-cookie": "^2.0.28",
-    "@types/react": "^0.14.57",
-    "@types/react-bootstrap": "0.0.37",
-    "@types/react-dom": "^0.14.23",
-    "@types/react-motion": "0.0.20",
-    "@types/react-redux": "^4.4.35",
-    "@types/react-router": "^2.0.41",
-    "@types/react-router-bootstrap": "0.0.27",
-    "@types/react-router-redux": "^4.0.35",
-    "@types/react-scroll": "0.0.30",
-    "@types/redux-thunk": "^2.1.28",
-    "@types/source-map": "^0.5.0",
-    "@types/uglify-js": "^2.6.28",
-    "@types/webpack": "^2.0.0",
-    "@types/webpack-env": "^1.12.1",
-    "@types/whatwg-fetch": "^0.0.33",
-    "animate-sass": "github:tgdev/animate-sass",
-    "aspnet-prerendering": "^2.0.3",
-    "aspnet-webpack": "^1.0.28",
-    "aspnet-webpack-react": "^1.0.5",
-    "awesome-typescript-loader": "^3.1.2",
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
-    "babel-plugin-lodash": "^3.2.11",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-3": "^6.22.0",
-    "babel-runtime": "^6.23.0",
-    "bootstrap-sass": "^3.3.7",
-    "classnames": "^2.2.5",
-    "compression-webpack-plugin": "^0.3.2",
-    "core-decorators": "^0.15.0",
-    "css-loader": "^0.26.1",
-    "domain-task": "^3.0.0",
-    "event-source-polyfill": "^0.0.9",
-    "extract-text-webpack-plugin": "^2.0.0-rc.3",
-    "file-loader": "^0.10.0",
-    "font-awesome": "^4.7.0",
-    "history": "^3.2.1",
-    "imports-loader": "^0.7.0",
-    "js-cookie": "^2.1.3",
-    "json-loader": "^0.5.4",
-    "lodash": "^4.17.4",
-    "lodash-webpack-plugin": "^0.11.0",
-    "node-noop": "^1.0.0",
-    "node-sass": "^4.5.0",
-    "postcss-cssnext": "^2.9.0",
-    "postcss-import": "^9.1.0",
-    "postcss-loader": "^1.2.2",
-    "preact": "^7.2.0",
-    "preact-compat": "^3.13.1",
-    "preact-dom": "^1.0.1",
-    "preact-render-to-string": "^3.6.0",
-    "precss": "^1.4.0",
-    "react": "^15.4.2",
-    "react-bootstrap": "^0.30.7",
-    "react-css-modules": "^4.1.0",
-    "react-dom": "^15.4.2",
-    "react-hot-loader": "^1.3.1",
-    "react-motion": "^0.4.7",
-    "react-redux": "^5.0.2",
-    "react-router": "^2.8.1",
-    "react-router-bootstrap": "^0.23.1",
-    "react-router-redux": "^4.0.7",
-    "react-scroll": "^1.4.5",
-    "redux": "^3.6.0",
-    "redux-thunk": "^2.2.0",
-    "resolve-url-loader": "^1.6.1",
-    "sass-loader": "^6.0.1",
-    "style-loader": "^0.13.1",
-    "tslib": "^1.6.0",
-    "ts-loader": "^2.0.0",
-    "typescript": "^2.2.1",
-    "uglify-js": "git+https://github.com/mishoo/UglifyJS2.git#harmony",
-    "url-loader": "^0.5.7",
-    "webpack": "^2.2.1",
-    "webpack-hot-middleware": "^2.17.0",
-    "webpack-merge": "^3.0.0"
-  },
-  "devDependencies": {
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-plugin-transform-regenerator": "^6.22.0"
+    "@types/classnames": "2.2.3",
+    "@types/history": "4.6.2",
+    "@types/js-cookie": "2.1.0",
+    "@types/lodash-es": "4.17.0",
+    "@types/msgpack5": "^3.4.1",
+    "@types/react": "^16.0.40",
+    "@types/react-bootstrap": "^0.32.8",
+    "@types/react-dom": "^16.0.5",
+    "@types/react-hot-loader": "^3.0.6",
+    "@types/react-redux": "^5.0.16",
+    "@types/react-router": "^4.0.23",
+    "@types/react-router-bootstrap": "0.24.4",
+    "@types/react-router-dom": "^4.0.8",
+    "@types/react-router-redux": "^5.0.13",
+    "@types/react-scroll": "1.5.3",
+    "@types/react-transition-group": "2.0.8",
+    "@types/webpack": "^4.1.3",
+    "@types/webpack-env": "1.13.6",
+    "ajv": "^6.2.1",
+    "animate-sass": "0.8.2",
+    "animated": "0.2.2",
+    "aspnet-prerendering": "3.0.1",
+    "aspnet-webpack": "2.0.3",
+    "aspnet-webpack-react": "3.0.0",
+    "awesome-typescript-loader": "5.0.0",
+    "bootstrap-sass": "3.3.7",
+    "buffer": "^5.1.0",
+    "caniuse-lite": "^1.0.30000821",
+    "classnames": "2.2.5",
+    "compression-webpack-plugin": "1.1.11",
+    "css-loader": "0.28.11",
+    "domain-task": "3.0.3",
+    "event-source-polyfill": "0.0.12",
+    "file-loader": "1.1.11",
+    "font-awesome": "4.7.0",
+    "history": "4.7.2",
+    "imports-loader": "0.8.0",
+    "isomorphic-fetch": "2.2.1",
+    "js-cookie": "2.2.0",
+    "json-loader": "0.5.7",
+    "lodash": "4.17.5",
+    "mini-css-extract-plugin": "^0.4.0",
+    "msgpack5": "^4.0.2",
+    "node-fetch": "2.1.1",
+    "node-noop": "1.0.0",
+    "node-sass": "4.8.3",
+    "postcss-cssnext": "3.1.0",
+    "postcss-import": "11.1.0",
+    "postcss-loader": "2.1.4",
+    "precss": "3.1.2",
+    "raw-loader": "0.5.1",
+    "react": "16.2.0",
+    "react-bootstrap": "0.32.1",
+    "react-css-modules": "4.7.2",
+    "react-dom": "16.2.0",
+    "react-hot-loader": "4.1.1",
+    "react-redux": "5.0.7",
+    "react-router": "4.2.0",
+    "react-router-bootstrap": "0.24.4",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "5.0.0-alpha.6",
+    "react-scroll": "1.6.1",
+    "react-transition-group": "2.3.1",
+    "redux": "4.0.0",
+    "redux-thunk": "2.2.0",
+    "resolve-url-loader": "2.3.0",
+    "sass-loader": "7.0.1",
+    "style-loader": "0.21.0",
+    "tslib": "1.9.0",
+    "typescript": "2.8.3",
+    "url-loader": "1.0.1",
+    "webpack": "4.6.0",
+    "webpack-bundle-analyzer": "2.11.1",
+    "webpack-cli": "^2.0.15",
+    "webpack-dev-middleware": "3.1.2",
+    "webpack-hot-middleware": "2.22.1",
+    "webpack-merge": "4.1.2"
   },
   "scripts": {
-    "build": "webpack --optimize-minimize"
+    "build:prod": "./node_modules/.bin/webpack --env.prod",
+    "build:debug": "./node_modules/.bin/webpack --display-error-details"
   }
 }
